Tables folder modularization into components, views and services
