+++ frontend/src/modules/Tables/components/index.js
@@ -1,0 +1,4 @@
+export * from './TableColumns';
+export * from './TableMetrics';
+export * from './TableOverview';
+export * from './TablePreview';
