+++ frontend/src/modules/Tables/services/index.js
@@ -1,0 +1,8 @@
+export * from './getDatasetTable';
+export * from './getDatasetTableProfilingRun';
+export * from './listDatasetTableProfilingRuns';
+export * from './previewTable';
+export * from './startDatasetProfilingRun';
+export * from './syncDatasetTableColumns';
+export * from './updateDatasetTable';
+export * from './updateDatasetTableColumn';
