+++ frontend/src/services/graphql/DatasetTable/index.js
@@ -1,11 +1,3 @@
 export * from './deleteDatasetTable';
-export * from './getDatasetTable';
-export * from './getDatasetTableProfilingRun';
 export * from './getSharedDatasetTables';
 export * from './listDatasetTableColumns';
-export * from './listDatasetTableProfilingRuns';
-export * from './previewTable';
-export * from './startProfilingRun';
-export * from './syncDatasetTableColumns';
-export * from './updateDatasetTable';
-export * from './updateDatasetTableColumn';
